refactor ROCAUCEvaluator
