+++ chainer_chemistry/training/extensions/__init__.py
@@ -1,4 +1,6 @@
+from chainer_chemistry.training.extensions import batch_evaluator  # NOQA
 from chainer_chemistry.training.extensions import roc_auc_evaluator  # NOQA
 
 # import class and function
+from chainer_chemistry.training.extensions.batch_evaluator import BatchEvaluator # NOQA
 from chainer_chemistry.training.extensions.roc_auc_evaluator import ROCAUCEvaluator  # NOQA
