+++ chainer_chemistry/training/extensions/roc_auc_evaluator.py
@@ -1,29 +1,9 @@
-import copy
-from logging import getLogger
-
 import numpy
 
-import chainer
-from chainer import cuda
 from chainer.dataset import convert
-from chainer import reporter
-from chainer.training.extensions import Evaluator
 from sklearn import metrics
 
-
-def _get_1d_numpy_array(v):
-    """Convert array or Variable to 1d numpy array
-
-    Args:
-        v (numpy.ndarray or cupy.ndarray or chainer.Variable): array to be
-            converted to 1d numpy array
-
-    Returns (numpy.ndarray): Raveled 1d numpy array
-
-    """
-    if isinstance(v, chainer.Variable):
-        v = v.data
-    return cuda.to_cpu(v).ravel()
+from chainer_chemistry.training.extensions.batch_evaluator import BatchEvaluator # NOQA
 
 
 def _to_list(a):
@@ -42,7 +22,7 @@ def _to_list(a):
         return a
 
 
-class ROCAUCEvaluator(Evaluator):
+class ROCAUCEvaluator(BatchEvaluator):
 
     """Evaluator which calculates ROC AUC score
 
@@ -95,42 +75,17 @@ class ROCAUCEvaluator(Evaluator):
                  logger=None):
         super(ROCAUCEvaluator, self).__init__(
             iterator, target, converter=converter, device=device,
-            eval_hook=eval_hook, eval_func=eval_func)
-        self.name = name
+            eval_hook=eval_hook, eval_func=eval_func, name=name,
+            logger=logger)
         self.pos_labels = _to_list(pos_labels)
         self.ignore_labels = _to_list(ignore_labels)
         self.raise_value_error = raise_value_error
-        self.logger = logger or getLogger()
 
-    def evaluate(self):
-        iterator = self._iterators['main']
-        eval_func = self.eval_func or self._targets['main']
+    @property
+    def metric_name(self):
+        return 'roc_auc'
 
-        if self.eval_hook:
-            self.eval_hook(self)
-
-        if hasattr(iterator, 'reset'):
-            iterator.reset()
-            it = iterator
-        else:
-            it = copy.copy(iterator)
-
-        y_total = []
-        t_total = []
-        for batch in it:
-            in_arrays = self.converter(batch, self.device)
-            with chainer.no_backprop_mode(), chainer.using_config('train',
-                                                                  False):
-                y = eval_func(*in_arrays[:-1])
-            t = in_arrays[-1]
-            y_data = _get_1d_numpy_array(y)
-            t_data = _get_1d_numpy_array(t)
-            y_total.append(y_data)
-            t_total.append(t_data)
-
-        y_total = numpy.concatenate(y_total).ravel()
-        t_total = numpy.concatenate(t_total).ravel()
-
+    def calc_metric(self, y_total, t_total):
         # --- ignore labels if specified ---
         if self.ignore_labels:
             valid_ind = numpy.in1d(t_total, self.ignore_labels, invert=True)
@@ -152,8 +107,4 @@ class ROCAUCEvaluator(Evaluator):
                     'ValueError detected during roc_auc_score calculation. {}'
                     .format(e.args))
                 roc_auc = numpy.nan
-
-        observation = {}
-        with reporter.report_scope(observation):
-            reporter.report({'roc_auc': roc_auc}, self._targets['main'])
-        return observation
+        return roc_auc
